[front] Refactor User api and fix validation
