--- front/src/containers/user/api/crud.js
@@ -1,5 +1,0 @@
-import { apiClient } from '../../../config/axios';
-
-export const getUser= async (id) => {
-  return apiClient.get(`/users/${id}`);
-}
