+++ front/src/containers/users/api/crud.js
@@ -3,3 +3,7 @@ import { apiClient } from '../../../config/axios';
 export const getUsers= async () => {
   return apiClient.get('/users');
 }
+
+export const getUser= async (id) => {
+  return apiClient.get(`/users/${id}`);
+}
