+++ front/src/containers/users/user/index.js
@@ -1,10 +1,10 @@
 import { useQuery } from 'react-query';
 import { useParams } from "react-router-dom";
 
-import { getUser } from './api/crud';
-import { Loading } from "../../components/loading";
-import { User } from "../../components/user";
-import { NotFound } from "../../components/404";
+import { getUser } from '../api/crud';
+import { Loading } from "../../../components/loading";
+import { User } from "../../../components/user";
+import { NotFound } from "../../../components/404";
 
 
 export const UserContainer = () => {
