+++ front/src/propsValidation/UserForListValidation.js
@@ -1,10 +1,8 @@
 const PropTypes = require("prop-types");
 
 module.exports = {
+  userId: PropTypes.number.isRequired,
   firstName: PropTypes.string.isRequired,
   lastName: PropTypes.string.isRequired,
-  email: PropTypes.string.isRequired,
   avatar: PropTypes.string,
-  phone: PropTypes.string.isRequired,
-  university: PropTypes.string
 }
